folder structure (added controllers)
