+++ app/middleware/authJwt.js
@@ -1,5 +1,6 @@
 const jwt = require("jsonwebtoken");
 
+// function checks jwt token
 verifyToken = (req, res, next) => {
   const authHeader = req.headers["authorization"];
   const token = authHeader && authHeader.split(' ')[1]
